Fixed regression in v1.1.0
The `edit` subcommand was broken by a regression that was introduced in
v1.1.0.

In `v1.1.0`, the `<id>` parameter of the `delete` subcommand was changed
such that it optionally accepted multiple parameters via the `...`
syntax in docopt. This unexpectedly changed the behavior of the `<id>`
parameter of the `edit` command as well (which did not and still does
not use the `...` syntax), which was causing its output (from docopt) to
be packaged in an array.

In turn, this array (rather than integer) was being passed as a
parameter to the prepapred statement in SQLite3. The query would always
run successfully, however, no records would ever be updated, because the
`<id>` value did not match any log entry primary keys.

The fix entailed simply addressing `<id>[0]` rather than `<id>`. This
change has been made.

This error wasn't detected by the unit tests initially, because the
`edit` subcommand would not call back when an error when no records were
updated. That behavior has now been changed, and unit-tests have been
written to monitor regressions in this new behavior.